+++ src/Modules/Meldeformular.php
@@ -125,42 +125,34 @@ class Meldeformular extends \Module
 			'inputType' => 'html',
 			'eval' => array
 			(
-				'html' => '<fieldset><legend>Turniere</legend><b>Hiermit melde ich mich zu folgenden Fernschachturnieren (bis zu 4 möglich in einem Formular):</b>'
+				'html' => '<fieldset><legend>Turnier</legend><b>Hiermit melde ich mich zu folgendem Fernschachturnier an:</b>'
 			)
 		)); 
-		for($x = 1; $x < 5;$x++)
-		{
-			$objForm->addFormField('turnierbox'.$x.'_start', array(
-				'inputType' => 'html',
-				'eval' => array
-				(
-					'html' => '<div style="display:flex;">'
-				)
-			)); 
-			$objForm->addFormField('turnier'.$x, array(
-				'label'         => 'Turnier',
-				'inputType'     => 'select',
-				'options'       => self::getTournaments(),
-				'eval'          => array('mandatory'=>false, 'choosen'=>true, 'includeBlankOption' => true)
-			));
-			$objForm->addFormField('anzahl'.$x, array(
-				'label'         => 'Anzahl',
-				'inputType'     => 'text',
-				'eval'          => array('mandatory'=>false)
-			));
-			$objForm->addFormField('nenngeld'.$x, array(
-				'label'         => 'Nenngeld in EUR',
-				'inputType'     => 'text',
-				'eval'          => array('mandatory'=>false,'rgxp'=> 'digit')
-			));
-			$objForm->addFormField('turnierbox'.$x.'_ende', array(
-				'inputType' => 'html',
-				'eval' => array
-				(
-					'html' => '</div>'
-				)
-			)); 
-		}
+		$objForm->addFormField('turnierbox_start', array(
+			'inputType' => 'html',
+			'eval' => array
+			(
+				'html' => '<div style="display:flex;">'
+			)
+		)); 
+		$objForm->addFormField('turnier', array(
+			'label'         => 'Turnier',
+			'inputType'     => 'select',
+			'options'       => self::getTournaments(),
+			'eval'          => array('mandatory'=>false, 'choosen'=>true, 'includeBlankOption' => true)
+		));
+		$objForm->addFormField('nenngeld', array(
+			'label'         => 'Nenngeld in EUR',
+			'inputType'     => 'text',
+			'eval'          => array('mandatory'=>false,'rgxp'=> 'digit')
+		));
+		$objForm->addFormField('turnierbox_ende', array(
+			'inputType' => 'html',
+			'eval' => array
+			(
+				'html' => '</div>'
+			)
+		)); 
 		$objForm->addFormField('fieldset2_ende', array(
 			'inputType' => 'html',
 			'eval' => array
@@ -179,7 +171,7 @@ class Meldeformular extends \Module
 		$objForm->addFormField('ueberweisung', array(
 			'label'         => 'wurde überwiesen am',
 			'inputType'     => 'text',
-			'eval'          => array('mandatory'=>false)
+			'eval'          => array('mandatory'=>false, 'rgxp'=>'date')
 		));
 		$objForm->addFormField('guthaben', array(
 			'label'         => array('', 'soll von meinem Guthaben beim BdF abgebucht werden'),
@@ -305,97 +297,136 @@ class Meldeformular extends \Module
 		// Datenbank aktualisieren
 		$zeit = time();
 		
-		// Turniere aus der Meldung der Reihe nach prüfen
-		for($x = 1; $x < 5; $x++)
+		// Turnier prüfen
+		if($data['turnier'])
 		{
-			if($data['turnier'.$x])
+			// Spieler in tl_fernschach_spieler suchen
+			$objSuche = \Database::getInstance()->prepare('SELECT id FROM tl_fernschach_spieler WHERE nachname=? AND vorname=? AND memberId=?')
+			                                    ->limit(1)
+			                                    ->execute($data['nachname'], $data['vorname'], $data['memberId']);
+			if($objSuche->numRows)
 			{
-				// Spieler in tl_fernschach_spieler suchen
-				$objSuche = \Database::getInstance()->prepare('SELECT id FROM tl_fernschach_spieler WHERE nachname=? AND vorname=? AND memberId=?')
-				                                    ->limit(1)
-				                                    ->execute($data['nachname'], $data['vorname'], $data['memberId']);
-				if($objSuche->numRows)
-				{
-					// Spieler gefunden
-					$spielerId = $objSuche->id;
-				}
-				else
-				{
-					$set = array
-					(
-						'tstamp'      => $zeit,
-						'nachname'    => $data['nachname'],
-						'vorname'     => $data['vorname'],
-						'plz'         => $data['plz'],
-						'ort'         => $data['ort'],
-						'strasse'     => $data['strasse'],
-						'email'       => $data['email'],
-						'memberId'    => $data['memberId'],
-					);
-					// Spieler nicht gefunden, dann neu eintragen
-					$objInsert = \Database::getInstance()->prepare('INSERT INTO tl_fernschach_spieler %s')
-					                                     ->set($set)
-					                                     ->execute();
-					$spielerId = $objInsert->insertId;
-				}
-				
-				// Anzahl Meldungen korrigieren
-				$data['anzahl'.$x] = $data['anzahl'.$x] ? (int)$data['anzahl'.$x] : 1;
-
-				for($y = 0; $y < $data['anzahl'.$x]; $y++)
-				{
-					$set = array
-					(
-						'tstamp'            => $zeit,
-						'spielerId'         => $spielerId,
-						'vorname'           => $data['vorname'],
-						'nachname'          => $data['nachname'],
-						'plz'               => $data['plz'],
-						'ort'               => $data['ort'],
-						'strasse'           => $data['strasse'],
-						'email'             => $data['email'],
-						'fax'               => $data['fax'],
-						'memberId'          => $data['memberId'],
-						'meldungDatum'      => $zeit,
-						'meldungTurnier'    => $data['turnier'.$x],
-						'meldungAnzahl'     => 1,
-						'meldungNenngeld'   => ($y == 0) ? $data['nenngeld'.$x] : '', // Nenngeld nur bei 1. Meldung eintragen
-						'nenngeldDate'      => $data['ueberweisung'],
-						'nenngeldGuthaben'  => $data['guthaben'],
-						'infoQualifikation' => $data['qualifikation'],
-						'bemerkungen'       => $data['bemerkungen'],
-						'published'         => 1
-					);
-					$objInsert = \Database::getInstance()->prepare('INSERT INTO tl_fernschach_meldungen %s')
-					                                     ->set($set)
-					                                     ->execute();
-				}
+				// Spieler gefunden
+				$spielerId = $objSuche->id;
 			}
+			else
+			{
+				$set = array
+				(
+					'tstamp'      => $zeit,
+					'nachname'    => $data['nachname'],
+					'vorname'     => $data['vorname'],
+					'plz'         => $data['plz'],
+					'ort'         => $data['ort'],
+					'strasse'     => $data['strasse'],
+					'email'       => $data['email'],
+					'memberId'    => $data['memberId'],
+				);
+				// Spieler nicht gefunden, dann neu eintragen
+				$objInsert = \Database::getInstance()->prepare('INSERT INTO tl_fernschach_spieler %s')
+				                                     ->set($set)
+				                                     ->execute();
+				$spielerId = $objInsert->insertId;
+			}
+			
+			$set = array
+			(
+				'pid'               => $data['turnier'],
+				'tstamp'            => $zeit,
+				'spielerId'         => $spielerId,
+				'vorname'           => $data['vorname'],
+				'nachname'          => $data['nachname'],
+				'plz'               => $data['plz'],
+				'ort'               => $data['ort'],
+				'strasse'           => $data['strasse'],
+				'email'             => $data['email'],
+				'fax'               => $data['fax'],
+				'memberId'          => $data['memberId'],
+				'meldungDatum'      => $zeit,
+				'meldungNenngeld'   => $data['nenngeld'],
+				'nenngeldDate'      => self::DatumToZeitstempel($data['ueberweisung']),
+				'nenngeldGuthaben'  => $data['guthaben'],
+				'infoQualifikation' => $data['qualifikation'],
+				'bemerkungen'       => $data['bemerkungen'],
+				'published'         => 1
+			);
+			//print_r($set);
+			$objInsert = \Database::getInstance()->prepare('INSERT INTO tl_fernschach_turniere_meldungen %s')
+			                                     ->set($set)
+			                                     ->execute();
 		}
 
         //
 		//\System::log('[Linkscollection] New Link submitted: '.$data['title'].' ('.$data['url'].')', __CLASS__.'::'.__FUNCTION__, TL_CRON);
         //
-		//// Email an Admin verschicken
-		//$objEmail = new \Email();
-		//$objEmail->from = $GLOBALS['TL_ADMIN_EMAIL'];
-		//$objEmail->fromName = $GLOBALS['TL_ADMIN_NAME'];
-		//$objEmail->subject = sprintf($GLOBALS['TL_LANG']['MSC']['linkscollection_subject'], \Idna::decode(\Environment::get('host')));
-        //
-		//// Kommentar zusammenbauen
-		//$strComment .= 'Titel: '.$data['title']."\n";
-		//$strComment .= 'URL: '.$data['url']."\n";
-		//$strComment .= 'Kategorie: '.$data['category']."\n";
-		//$strComment .= 'Beschreibung: '.$data['description'];
-        //
-		//// Add the comment details
+		$objTurnier = \Schachbulle\ContaoFernschachBundle\Classes\Helper::getTurnierdatensatz($data['turnier']);
+		
+		// E-Mail für Turnierleiter zusammenbauen
+		$turnierleiter = self::getTurnierleiter($data['turnier']);
+		
+		if(isset($turnierleiter[0]))
+		{
+			// Email verschicken
+			$objEmail = new \Email();
+			$objEmail->charset = 'utf-8';
+			$objEmail->from = $GLOBALS['TL_CONFIG']['fernschach_emailAdresse'];
+			$objEmail->fromName = $GLOBALS['TL_CONFIG']['fernschach_emailVon'];
+			$objEmail->subject = 'Turnieranmeldung '.$objTurnier->title;
+			$objEmail->replyTo($turnierleiter[0]['name'].' <'.$turnierleiter[0]['email'].'>');
+			// Weitere Empfänger einbauen
+			if(count($turnierleiter) > 1)
+			{
+				$empfaenger = array();
+				for($x = 1; $x < count($turnierleiter); $x++)
+				{
+					if($turnierleiter[$x]['email']) $empfaenger[] = $turnierleiter[$x]['name'] . ' <' . $turnierleiter[$x]['email'] . '>';
+				}
+				$cc = implode(',', $empfaenger);
+				$objEmail->sendCc($cc);  
+			}
+			// Kommentar zusammenbauen
+			$text = '<html><head><title></title></head><body>';
+			$text .= '<p>Eine neue Turnieranmeldung wurde vorgenommen:</p>';
+			$text .= '<h3>Angaben zum Turnier</h3>';
+			$text .= '<ul>';
+			$text .= '<li>Turnier: <b>'.$objTurnier->title.'</b></li>';
+			$text .= '<li>Meldeschluss: <b>'.($objTurnier->registrationDate ? date('d.m.Y', $objTurnier->registrationDate) : '-').'</b></li>';
+			$text .= '</ul>';
+			$text .= '<h3>Angaben zum Spieler</h3>';
+			$text .= '<ul>';
+			$text .= '<li>Datum und Uhrzeit: <b>'.date('d.m.Y H:i', $zeit).'</b></li>';
+			$text .= '<li>Vor- und Nachname: <b>'.$data['vorname'].' '.$data['nachname'].'</b></li>';
+			$text .= '<li>BdF_mitgliedsnummer: <b>'.$data['memberId'].'</b></li>';
+			$text .= '<li>Adresse: <b>'.$data['plz'].' '.$data['ort'].', '.$data['strasse'].'</b></li>';
+			$text .= '<li>Fax: <b>'.$data['fax'].'</b></li>';
+			$text .= '<li>E-Mail: <b>'.$data['email'].'</b></li>';
+			$text .= '</ul>';
+			$text .= '<h3>Angaben zum Nenngeld</h3>';
+			$text .= '<ul>';
+			$text .= '<li>Nenngeld: <b>'.$data['nenngeld'].'</b></li>';
+			$text .= '<li>Überwiesen am: <b>'.$data['ueberweisung'].'</b></li>';
+			$text .= '<li>Abbuchen vom Guthaben: <b>'.$data['guthaben'].'</b></li>';
+			$text .= '</ul>';
+			$text .= '<h3>Sonstiges</h3>';
+			$text .= '<ul>';
+			$text .= '<li>Information zur Qualifikation: <b>'.$data['qualifikation'].'</b></li>';
+			$text .= '<li>Bemerkungen: <b>'.$data['bemerkungen'].'</b></li>';
+			$text .= '</ul>';
+			$text .= '<p><i>Diese E-Mail wurde automatisch erstellt.</i></p></body></html>';
+			//	'pid'               => $data['turnier'],
+			//	'tstamp'            => $zeit,
+			//	'spielerId'         => $spielerId,
+        	
+			// Add the comment details
+			$objEmail->html = $text;
+			$objEmail->sendTo(array($turnierleiter[0]['name'].' <'.$turnierleiter[0]['email'].'>'));
+		}
 		//$objEmail->text = sprintf($GLOBALS['TL_LANG']['MSC']['linkscollection_message'],
 		//                          $data['name'] . ' (' . $data['email'] . ')',
 		//                          $strComment,
 		//                          \Idna::decode(\Environment::get('base')) . \Environment::get('request'),
 		//                          \Idna::decode(\Environment::get('base')) . 'contao/main.php?do=linkscollection&table=tl_linkscollection_links&act=edit&id=' . $objLink->insertId);
-        //
-		//$objEmail->sendTo(array($GLOBALS['TL_ADMIN_NAME'].' <'.$GLOBALS['TL_ADMIN_EMAIL'].'>'));
+        
 	}
 
 	/**
@@ -408,17 +439,71 @@ class Meldeformular extends \Module
 	public function getTournaments()
 	{
 		$arrForms = array();
-		$objForms = $this->Database->prepare("SELECT * FROM tl_fernschach_turniere WHERE (registrationDate > ? OR registrationDate = ?) AND onlineAnmeldung = ? AND published = ? ORDER BY art AND title")
-		                           ->execute(time(), 0, 1, 1);
+		$objTurniere = \Database::getInstance()->prepare("SELECT * FROM tl_fernschach_turniere WHERE (registrationDate > ? OR registrationDate = ?) AND onlineAnmeldung = ? AND published = ? ORDER BY art AND title")
+		                                       ->execute(time(), 0, 1, 1);
 
-		while ($objForms->next())
+		while($objTurniere->next())
 		{
-			$meldedatum = $objForms->registrationDate ? '  | Meldedatum: '.date('d.m.Y', $objForms->registrationDate) : '  | ohne Meldedatum';
-			$nenngeld = ' | Nenngeld: '.str_replace('.', ',', sprintf('%0.2f', $objForms->nenngeld)).' € ';
-			$arrForms[$objForms->id] = $objForms->title.$nenngeld.$meldedatum;
+			$published = self::TurnierkategorieVeroeffentlicht($objTurniere->pid); // Prüfen, ob alle übergeordneten Turnierkategorien veröffentlicht sind
+			
+			if($published)
+			{
+				$meldedatum = $objTurniere->registrationDate ? '  | Meldedatum: '.date('d.m.Y', $objForms->registrationDate) : '  | ohne Meldedatum';
+				$nenngeld = ' | Nenngeld: '.str_replace('.', ',', sprintf('%0.2f', $objTurniere->nenngeld)).' € ';
+				$arrForms[$objTurniere->id] = $objTurniere->title.$nenngeld.$meldedatum;
+			}
 		}
 
 		return $arrForms;
+	}
+
+	/*
+	 * Funktion TurnierkategorieVeroeffentlicht
+	 * Liefert true/false, je nach veröffentlichten Oberkategorien
+	 */
+	private function TurnierkategorieVeroeffentlicht($id)
+	{
+		while($id > 0)
+		{
+			$objTurnier = \Database::getInstance()->prepare("SELECT * FROM tl_fernschach_turniere WHERE id = ?")
+			                                      ->execute($id);
+
+			if($objTurnier->published) 
+			{
+				$id = $objTurnier->pid; // Neue ID setzen
+			}
+			else
+			{
+				return false; // Kategorie ist nicht veröffentlicht
+			}
+		}
+		return true;
+	}
+
+	/*
+	 * Funktion getTurnierleiter
+	 * Liefert die Turnierleiter (Name und E-Mail) eines Turniers und seiner Oberkategorien
+	 */
+	private function getTurnierleiter($id)
+	{
+		$arr = array();
+		while($id > 0)
+		{
+			$objTurnier = \Database::getInstance()->prepare("SELECT * FROM tl_fernschach_turniere WHERE id = ?")
+			                                      ->execute($id);
+
+			if($objTurnier->published && $objTurnier->turnierleiterInfo && $objTurnier->turnierleiterEmail) 
+			{
+				// Turnierleiter speichern
+				$arr[] = array
+				(
+					'name'    => $objTurnier->turnierleiterName,
+					'email'   => $objTurnier->turnierleiterEmail
+				);
+			}
+			$id = $objTurnier->pid; // Neue ID setzen
+		}
+		return $arr;
 	}
 
 	/*
@@ -442,6 +527,18 @@ class Meldeformular extends \Module
 		if (null !== $objNotification) {
 			$objNotification->send($arrTokens);
 		}
+	}
+
+	/*
+	 * Funktion DatumToZeitstempel
+	 * Datum TT.MM.JJJJ in Unix-Timestamp umwandeln
+	 */
+	private function DatumToZeitstempel($string)
+	{
+		$tag = substr($string, 0, 2);
+		$monat = substr($string, 3, 2);
+		$jahr = substr($string, 6, 4);
+		return mktime(0, 0, 0, $monat, $tag, $jahr);
 	}
 
 }