+++ src/Resources/contao/dca/tl_fernschach_turniere_meldungen.php
@@ -265,10 +265,10 @@ $GLOBALS['TL_DCA']['tl_fernschach_turniere_meldungen'] = array
 				'datepicker'          => true,
 				'tl_class'            => 'w50 wizard'
 			),
-			'load_callback' => array
-			(
-				array('tl_fernschach_turniere_meldungen', 'loadDate')
-			),
+			//'load_callback' => array
+			//(
+			//	array('tl_fernschach_turniere_meldungen', 'loadDate')
+			//),
 			'sql'                     => "int(10) unsigned NOT NULL default 0"
 		),
 		'nenngeldInfo' => array
@@ -415,10 +415,17 @@ class tl_fernschach_turniere_meldungen extends \Backend
 		$temp = '<div class="tl_content_left">';
 
 		// Vor- und Nachname
-		if($arrRow['state'] == 0) $temp .= '<b>'.$arrRow['vorname'].' '.$arrRow['nachname'].'</b>';
-		elseif($arrRow['state'] == 1) $temp .= '<b style="color:green">'.$arrRow['vorname'].' '.$arrRow['nachname'].'</b>';
-		else $temp .= '<b style="color:red">'.$arrRow['vorname'].' '.$arrRow['nachname'].'</b>';
-
+		if(isset($arrRow['state']))
+		{
+			if($arrRow['state'] == 0) $temp .= '<b>'.$arrRow['vorname'].' '.$arrRow['nachname'].'</b>';
+			elseif($arrRow['state'] == 1) $temp .= '<b style="color:green">'.$arrRow['vorname'].' '.$arrRow['nachname'].'</b>';
+			else $temp .= '<b style="color:red">'.$arrRow['vorname'].' '.$arrRow['nachname'].'</b>';
+		}
+		else
+		{
+			$temp .= '<b>'.$arrRow['vorname'].' '.$arrRow['nachname'].'</b>';
+		}
+		
 		// Zuordnung
 		if($arrRow['spielerId'])
 		{